+++ compost/event.ts
@@ -1,0 +1,3 @@
+import { modinfo } from "../modinfo.ts";
+
+export const COMPOST_SETTLE_EVENT = `${modinfo.id}:compost:settle`;